// src/app/interfaces/Collaborators.ts
export interface Collaborators {
  id: number;
  contratType: string;
  dureeTravail: string;
  dateDebutContrat: Date;
  dateFinContrat: Date;
  anciennete: number;
  categorieSocioProfessionnelle: string;
  departement: string;
  genre: string;
  age: number;
  niveauEducation: string;
  regionGeographique: string;
  populationName: string ;
  collaborator: any;


}