// src/app/interfaces/CollaboratorsAdd.ts
export interface CollaboratorsAdd {
  id: number;
  contratType: string;
  dureeTravail: string;
  dateDebutContrat: Date;
  dateFinContrat: Date;
  anciennete: number;
  categorieSocioProfessionnelle: string;
  departement: string;
  genre: string;
  age: number;
  niveauEducation: string;
  regionGeographique: string;
  populationName: string;
  firstname:string;
  lastname:string;
  email:string;
  password:number;
  collaborator: any;

}